+++ packages/mask/src/masks.ts
@@ -15,6 +15,7 @@ export const masks = {
   trackNumber: 'AA999999999AA',
   monthAndYear: '99/9999',
   date: '99/99/9999',
+  daterange: '99/99/9999 - 99/99/9999',
 }
 
 export type MaskType = keyof typeof masks