+++ packages/mask/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @siakit/mask
 
+## 0.0.7
+
+### Patch Changes
+
+- add date range picker
+
 ## 0.0.6
 
 ### Patch Changes